backend/Merges/requirements.txt
pandas
python-docx
python-dotenv
msal
requests
openpyxl